[refact] 우승자 추출 메서드 이동
RacingCars가 가지고 있던 우승자를 추출하는 로직을
Winners로 이동
